+++ src/main/kotlin/space/kscience/gradle/KScienceReadmeExtension.kt
@@ -9,10 +9,7 @@ import freemarker.template.TemplateNotFoundException
 import kotlinx.html.TagConsumer
 import kotlinx.html.div
 import kotlinx.html.stream.createHTML
-import kotlinx.serialization.Serializable
-import kotlinx.serialization.json.Json
 import org.gradle.api.Project
-import org.gradle.internal.cc.base.logger
 import org.gradle.kotlin.dsl.*
 import org.intellij.lang.annotations.Language
 import org.jetbrains.dokka.gradle.tasks.DokkaGenerateTask
@@ -58,8 +55,6 @@ public class KScienceReadmeExtension(private val kscience: KSciencePlatformExten
             }
         }
 
-    public var featuresFile: File = project.file("docs/features.json")
-
     private val fmLoader = StringTemplateLoader().apply {
         putTemplate(
             "artifact",
@@ -80,29 +75,25 @@ public class KScienceReadmeExtension(private val kscience: KSciencePlatformExten
         templateLoader = fmLoader
     }
 
-    @Serializable
-    public data class Feature(val id: String, val description: String, val ref: String?, val name: String = id)
+    public data class Feature(val id: String, val description: String, val ref: String?, val name: String = id): java.io.Serializable
 
-    @Deprecated("Put features into docs/features.json instead")
-    public val manualFeatures: MutableList<Feature> = mutableListOf()
+    public val features: MutableList<Feature> = mutableListOf()
 
     /**
      * A plain readme feature with description
      */
-    @Deprecated("Put features into docs/features.json instead")
     public fun feature(
         id: String,
         @Language("File") ref: String? = null,
         name: String = id,
         @Language("markdown") description: () -> String,
     ) {
-        manualFeatures += Feature(id, description(), ref, name)
+        features += Feature(id, description(), ref, name)
     }
 
     /**
      * A readme feature with HTML description
      */
-    @Deprecated("Put features into docs/features.json instead")
     public fun featureWithHtml(
         id: String,
         ref: String? = null,
@@ -114,7 +105,7 @@ public class KScienceReadmeExtension(private val kscience: KSciencePlatformExten
                 htmlBuilder()
             }
         }.finalize()
-        manualFeatures += Feature(id, text, ref, name)
+        features += Feature(id, text, ref, name)
     }
 
     private val properties: MutableMap<String, Project.() -> Any?> = mutableMapOf(
@@ -190,15 +181,6 @@ public class KScienceReadmeExtension(private val kscience: KSciencePlatformExten
      * Generate a Markdown string listing features
      */
     internal fun featuresString(itemPrefix: String = " - ", pathPrefix: String = ""): String = buildString {
-        val features = if (featuresFile.exists()) {
-            json.decodeFromString<List<Feature>>(featuresFile.readText())
-        } else {
-            manualFeatures.also {
-                logger.info("No features file found, using manual features and writing them to ${featuresFile.absolutePath}")
-                featuresFile.writeText(json.encodeToString(manualFeatures))
-            }
-        }
-
         features.forEach {
             appendLine("$itemPrefix[${it.name}]($pathPrefix${it.ref ?: "#"}) : ${it.description}")
         }
@@ -226,7 +208,7 @@ public class KScienceReadmeExtension(private val kscience: KSciencePlatformExten
         if (readmeTemplate != other.readmeTemplate) return false
         if (fmLoader != other.fmLoader) return false
         if (fmCfg != other.fmCfg) return false
-        if (manualFeatures != other.manualFeatures) return false
+        if (features != other.features) return false
         if (properties != other.properties) return false
         if (inputFiles != other.inputFiles) return false
 
@@ -240,17 +222,13 @@ public class KScienceReadmeExtension(private val kscience: KSciencePlatformExten
         result = 31 * result + readmeTemplate.hashCode()
         result = 31 * result + fmLoader.hashCode()
         result = 31 * result + fmCfg.hashCode()
-        result = 31 * result + manualFeatures.hashCode()
+        result = 31 * result + features.hashCode()
         result = 31 * result + properties.hashCode()
         result = 31 * result + inputFiles.hashCode()
         return result
     }
 
     public companion object{
-        public val json: Json = Json {
-            prettyPrint = true
-            ignoreUnknownKeys = true
-        }
     }
 
 }
@@ -272,11 +250,10 @@ internal fun KSciencePlatformExtension.configureReadme() = with(project) {
         group = "documentation"
         description = "Generate a README file if stub is present"
 
-        inputs.property("features", readmeExtension.manualFeatures)
+        inputs.property("features", readmeExtension.features)
 
         if (readmeExtension.readmeTemplate.exists()) {
             inputs.file(readmeExtension.readmeTemplate)
-            inputs.file(readmeExtension.featuresFile)
         }
 
         readmeExtension.inputFiles.forEach {
@@ -292,7 +269,7 @@ internal fun KSciencePlatformExtension.configureReadme() = with(project) {
                     tasks.findByName("generateReadme")?.let { readmeTask ->
                         dependsOn(readmeTask)
                     }
-                    inputs.property("features-${name}", subProjectReadmeExtension.manualFeatures)
+                    inputs.property("features-${name}", subProjectReadmeExtension.features)
                 }
             }
         }